Improve the advanced file downloader
It's a complete demo for file downloading (which I even have been using recently to download some files) so I added a progress bar (which many people have issues implementing) and a --verbose option. Also moved the file verification to its own function so people can reuse it if they desire. And sorted files because it looks nicer.